+++ protocols/kad/src/lib.rs
@@ -38,6 +38,7 @@
 // TODO: we allow dead_code for now because this library contains a lot of unused code that will
 //       be useful later for record store
 #![allow(dead_code)]
+#![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 
 pub mod handler;
 pub mod kbucket;